FIX: 가족감정조회시그룹이름조회추가
