+++ src/main/java/da_ni_ni/backend/emotion/dto/FindTotalEmotionsResponse.java
@@ -3,6 +3,7 @@ package da_ni_ni.backend.emotion.dto;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import da_ni_ni.backend.common.ResponseDto;
+import da_ni_ni.backend.group.domain.FamilyGroup;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -12,10 +13,13 @@ import java.util.List;
 @Builder
 @JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class FindTotalEmotionsResponse implements ResponseDto {
+    private String groupName;
     private List<FindEmotionDetailResponse> emotionList;
 
-    public static FindTotalEmotionsResponse createWith(List<FindEmotionDetailResponse> emotionList) {
+    public static FindTotalEmotionsResponse createWith(FamilyGroup familyGroup
+                                                       , List<FindEmotionDetailResponse> emotionList) {
         return FindTotalEmotionsResponse.builder()
+                .groupName(familyGroup.getName())
                 .emotionList(emotionList)
                 .build();
     }
