+++ src/main/java/da_ni_ni/backend/group/service/GroupService.java
@@ -124,19 +124,4 @@ public class GroupService {
         return ApprovejoinResponse.createWith(targetRequest);
     }
 
-    // 가족명 수정 (O)
-    public UpdateGroupNameResponse updateGroupName(Long userId, UpdateGroupNameRequest request) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(UserNotFoundException::new);
-        FamilyGroup familyGroup = user.getFamilyGroup();
-        if (familyGroup == null) {
-            throw new GroupNotFoundException();
-        }
-        UpdateGroupNameData updateGroupNameData = UpdateGroupNameData.createWith(request);
-        familyGroup.updateName(updateGroupNameData);
-        groupRepository.save(familyGroup);
-        return UpdateGroupNameResponse.createWith(familyGroup);
-    }
-
-
 }
